Remove special collections from random generation
There are some collections with special needs like Comparable items or synchronicity requirements. These can't be used in the usual Collection::add / Collection::iterator way and will throw exceptions when trying to interact with them in those ways.

To prevent this, we removed such collections from the list of collections that can be randomly generated.

This included:

* `ConcurrentSkipListSet` (natural ordering, elements must implement Comparable)
* `PriorityBlockingQueue` (natural ordering)
* `SynchronousQueue` (capacity 0, elements must be synchronously added via Queue::offer)
* `TreeSet` (natural ordering)

Resolves #23

Signed-off-by: Nirro <nireiorange@gmail.com>